requirements.txt
alembic==1.13.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.0
billiard==4.2.1
boto3==1.35.60
botocore==1.35.60
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
olefile==0.47
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==5.2.0
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
